+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -4,6 +4,8 @@ object Versions {
     const val androidPlugin = "3.3.0"
     const val androidXCollection = "1.0.0"
     const val androidXFragment = "1.0.0"
+    const val androidXLifecycleExtensions = "2.0.0"
+    const val androidXLifecycleViewModel = "2.0.0"
     const val bintrayPlugin = "1.8.4"
     const val dokkaPlugin = "0.9.17"
     const val kluent = "1.47"
@@ -17,6 +19,10 @@ object Dependencies {
     const val androidPlugin = "com.android.tools.build:gradle:${Versions.androidPlugin}"
     const val androidXCollection = "androidx.collection:collection:${Versions.androidXCollection}"
     const val androidXFragment = "androidx.fragment:fragment:${Versions.androidXFragment}"
+    const val androidXLifecycleExtensions =
+        "androidx.lifecycle:lifecycle-extensions:${Versions.androidXLifecycleExtensions}"
+    const val androidXLifecycleViewModel =
+        "androidx.lifecycle:lifecycle-viewmodel:${Versions.androidXLifecycleViewModel}"
     const val kluent = "org.amshove.kluent:kluent:${Versions.kluent}"
     const val spek2RunnerJunit5 = "org.spekframework.spek2:spek-runner-junit5:${Versions.spek}"
     const val spekApi = "org.spekframework.spek2:spek-dsl-jvm:${Versions.spek}"